api/requirements.txt
camelot-py==0.10.1
chardet==5.1.0
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
dnspython==2.2.1
email-validator==1.3.0
et-xmlfile==1.1.0
fastapi==0.88.0
fonttools==4.38.0
ghostscript==0.7
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
itsdangerous==2.1.2
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.24.1
opencv-python==4.7.0.68
openpyxl==3.0.10
orjson==3.8.4
pandas==1.5.2
pdfminer.six==20221105
pdftopng==0.2.3
Pillow==9.4.0
pydantic==1.10.4
pyparsing==3.0.9
PyPDF2==2.12.1
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.7
rfc3986==1.5.0
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
starlette==0.22.0
tabulate==0.9.0
threadpoolctl==3.1.0
ujson==5.7.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
xgboost==1.7.3
